backend/requirements.txt
# Core Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22

# Configuration Management
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.0

# HTTP Client & API Communication
httpx==0.27.0

# Image Processing
Pillow==12.1.1

# Utilities
cachetools==5.3.3
sse-starlette==2.1.0
tenacity==8.2.3
